Fix function prototypes in languages.h to match code generated by gperf

+++ languages.h
@@ -68,8 +68,8 @@ typedef struct {
     Language language;
 } LanguageHashSlot;
 
-const LanguageHashSlot *lookup_language_by_extension(const char*, unsigned int);
-const LanguageHashSlot *lookup_language_by_filename(const char*, unsigned int);
+const LanguageHashSlot *lookup_language_by_extension(const char*, size_t);
+const LanguageHashSlot *lookup_language_by_filename(const char*, size_t);
 const char *lookup_language_name(Language lang);
 Parser lookup_language_parser(Language);
 Language detect_language(const char *path, int base, int level, size_t size);
